==> fishing/R Scripts/0.2 building the focal gears dataframe.R <==
rm(list = ls()) #fresh start
#setwd("output3\\discards")
library(ggplot2)
library(tidyverse)
library(furrr)
library(tictoc)
library(gtools)
library(gsubfn)

tic("Mapping the dataframe") #time
plan("multisession") #setup parallel processing

scaling <- seq(0,3.99,0.01) #define scale factors from focal gear file
files <- list.files() #lists files - need to set wd first
files2 <- mixedsort(sort(files)) #sort the files
master <- read.csv("..\\..\\mastercombinationsdf.csv") %>% 
  select(-1) #load in master dataframe and delete the row index

result_lst <- future_map(.x = files2,.f = readRDS,.progress = TRUE) %>% #apply readRDS to files in parallel
  lapply(function(x) as.data.frame(x)) %>% #transform into df
  mapply(`[<-`, ., 'SimulationID', value = master$SimulationID,
         SIMPLIFY = FALSE) %>% #adds simulation ID to each dataframe
  mapply(`[<-`, ., 'Zone', value = rep(c("Inshore","Offshore"),length(files2)/2),
         SIMPLIFY = FALSE) %>% #adds inshore/offshore columns
  mapply(`[<-`, ., 'Focal Gear', value = gsub(".*shore_","",files2),
         SIMPLIFY = FALSE) %>% #adds filenames
  mapply(`[<-`, ., 'Focal Multiplier', value = rep(scaling,each = 2),
         SIMPLIFY = FALSE)#adds focal multiplier
toc()
tic("Building the dataframe")
finaldiscards <- data.frame() #empty df
investigating <- c("Pelagic invertebrates",
                   "Birds","Quota-limited demersal fish","Carn/scavenge feeding benthos" ,"Cetaceans" ,"Non-quota demersal fish" ,
                   "Migratory fish","Susp/deposit feeding benthos","Planktivorous fish","Macrophytes","Pinnipeds")
for(i in 1:length(investigating)){
  guild <- lapply(result_lst, function(result_lst) subset(result_lst, row.names(result_lst) == investigating[i]))  #subsets to just one guild
  
  result <- do.call(rbind,guild) %>% #adds guilds
    data.frame() #turns in to a dataframe
  result <- cbind(Guild = investigating[i],result)
  row.names(result) <- NULL #reset rownames
  
  result$Total <- rowSums(result[,3:13]) #calculates total
  
  finaldiscards <- rbind(finaldiscards,result) #bind all together
  print(paste0(i,"/",length(investigating))) #progress
}
row.names(finaldiscards) <- NULL #reset rownames

finaldiscards$Focal.Gear<-gsub(".rds","",as.character(finaldiscards$Focal.Gear)) #fix gearnames
toc()
cols <- colnames(finaldiscards)[2:13]
finaldiscardsgears <- pivot_longer(data = finaldiscards,cols = cols,
                      names_to = "Gear",
                      values_to = "Catch")

toc()
rm(cols,files,files2,i,investigating,scaling,result_lst,result,master,guild)
